chapter_001/src/test/java/ru/job4j/loop/CheckPrimeNumberTest.java
package ru.job4j.loop;

import org.junit.Test;

import static org.hamcrest.Matchers.is;
import static org.junit.Assert.*;

public class CheckPrimeNumberTest {
    @Test
    public void when5() {
        boolean rsl = CheckPrimeNumber.check(5);
        assertThat(rsl, is(true));
    }

    @Test
    public void when4() {
        boolean rsl = CheckPrimeNumber.check(4);
        assertThat(rsl, is(false));
    }

    @Test
    public void when1() {
        boolean rsl = CheckPrimeNumber.check(1);
        assertThat(rsl, is(false));
    }

    @Test
    public void when12() {
        boolean rsl = CheckPrimeNumber.check(12);
        assertThat(rsl, is(false));
    }

    @Test
    public void when269() {
        boolean rsl = CheckPrimeNumber.check(269);
        assertThat(rsl, is(true));

    }

    @Test
    public void when268() {
        boolean rsl = CheckPrimeNumber.check(268);
        assertThat(rsl, is(false));
    }

}